board/src/main/java/study/board/service/UserService.java
package study.board.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import study.board.domain.User;
import study.board.repository.UserRepository;

import java.util.List;
import java.util.Optional;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class UserService {

    private final UserRepository userRepository;

    /**
     * 회원가입
     */
    @Transactional
    public Long register(User user) {
        validationId(user);
        userRepository.save(user);
        return user.getId();
    }

    private void validationId(User user) {
        userRepository.findByLoginId(user.getLoginId())
                .ifPresent(e -> {
                    throw new IllegalStateException("이미 존재하는 아이디입니다.");
                });
    }

    /**
     * 회원 조회
     */
    public Optional<User> findById(Long id) {
        return userRepository.findById(id);
    }

    /**
     * 회원 전체 조회
     */
    public List<User> findUsers() {
        return userRepository.findAll();
    }
}
